refactor: 테스트만을 위한 메소드 삭제, 패키지 추가 등
